switch minHeap value to uint in all instances to avoid sign errors

+++ app/application.cpp
@@ -339,9 +339,9 @@ size_t Application::getFreeHeapSize(){
 	return fh;
 }
 
-bool Application::checkHeap( size_t minHeap)
+bool Application::checkHeap( uint minHeap)
 {
-	int fh = getFreeHeapSize();
+	uint fh = getFreeHeapSize();
 	if (fh<6000)
 	{
 		// minimize heap usage by increasing the minHeap threshold when we're critical anyway. This should preserve some heap for receive packet buffers and thus improve stability
